Add getProposedNonce
